+++ extfile/extfile.go
@@ -1,0 +1,38 @@
+package extfile
+
+import (
+	"bytes"
+	"encoding/base64"
+	"io"
+	"os"
+)
+
+func AppendToFile(filename string, lines []string) error {
+	f, err := os.OpenFile(filename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+	for _, line := range lines {
+		if _, err := f.WriteString(line); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func File2Base64(file string) (string, error) {
+	f, err := os.Open(file)
+	if err != nil {
+		return "", err
+	}
+	defer f.Close()
+
+	var buffer bytes.Buffer
+	_, err = io.Copy(&buffer, f)
+	if err != nil {
+		return "", err
+	}
+
+	return base64.StdEncoding.EncodeToString(buffer.Bytes()), nil
+}